BCB: Added logging SEH information: exception code & address. Core/Engine: Added GetUnreadLog method.

+++ Core/Engine/UEngine.h
@@ -114,7 +114,7 @@ mutable string TempLogString;
 mutable string CompName;
 
 // Индекс последнего считанного символа лога исключений
-size_t LastReadExceptionLogIndex;
+int LastReadExceptionLogIndex;
 
 public: // Методы
 // --------------------------
